Revert "Feature/frontend add list and filtering"
