+++ resq/frontend/package.json
@@ -11,13 +11,11 @@
     "@mui/material": "^5.14.14",
     "@mui/styled-engine": "^5.14.14",
     "@mui/x-data-grid": "^6.18.1",
-    "@mui/x-date-pickers": "^6.18.2",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.6.0",
     "bootstrap": "^5.3.2",
-    "dayjs": "^1.11.10",
     "next": "^14.0.0",
     "npm": "^10.2.1",
     "pigeon-maps": "^0.21.3",
