+++ resq/frontend/src/App.js
@@ -9,8 +9,6 @@ import Account from "./pages/Account";
 import RoleRequest from "./pages/RoleRequest";
 import LogoutIcon from '@mui/icons-material/Logout';
 import Request from "./pages/RequestCreation";
-import {LocalizationProvider} from "@mui/x-date-pickers";
-import {AdapterDayjs} from "@mui/x-date-pickers/AdapterDayjs";
 
 const SmallRedCircle = () =>
     <svg
@@ -25,8 +23,11 @@ const SmallRedCircle = () =>
 function App() {
     const [token, _setToken] = useState(localStorage.getItem("token"))
     const [role, setRole] = useState("")
+    // eslint-disable-next-line no-unused-vars
+    const [width, setWidth]   = useState(window.innerWidth);
     const [height, setHeight] = useState(window.innerHeight);
     const updateDimensions = () => {
+        setWidth(window.innerWidth);
         setHeight(window.innerHeight);
     }
     useEffect(() => {
@@ -62,83 +63,81 @@ function App() {
     const ref = useRef(null)
 
     return (
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-            <Router>
-                <div>
-                    <Navbar bg="light" variant="light" expand="lg">
-                        <Container ref={ref}>
-                            <Navbar.Brand href="/" style={{color: 'red', fontWeight: 'bold'}}>
-                                <SmallRedCircle/>
-                                ResQ
-                            </Navbar.Brand>
-                            <Navbar.Toggle aria-controls="basic-navbar-nav"/>
-                            <Navbar.Collapse id="basic-navbar-nav">
-                                <Nav className="me-auto">
-                                    {navLinks.map(({path, label, icon}) => (
-                                        <Nav.Link key={path} href={path}>
-                                            {icon}
-                                            {label}
+        <Router>
+            <div>
+                <Navbar bg="light" variant="light" expand="lg">
+                    <Container ref={ref}>
+                        <Navbar.Brand href="/" style={{color: 'red', fontWeight: 'bold'}}>
+                            <SmallRedCircle/>
+                            ResQ
+                        </Navbar.Brand>
+                        <Navbar.Toggle aria-controls="basic-navbar-nav"/>
+                        <Navbar.Collapse id="basic-navbar-nav">
+                            <Nav className="me-auto">
+                                {navLinks.map(({path, label, icon}) => (
+                                    <Nav.Link key={path} href={path}>
+                                        {icon}
+                                        {label}
+                                    </Nav.Link>
+                                ))}
+                            </Nav>
+                            <Nav className="ml-auto">
+                                {token ?
+                                    <>
+                                        <Nav.Link key={"/account"} href={"/account"} style={{"marginLeft": "auto"}}>
+                                            <SmallRedCircle/>
+                                            <strong>Account</strong>
                                         </Nav.Link>
-                                    ))}
-                                </Nav>
-                                <Nav className="ml-auto">
-                                    {token ?
-                                        <>
-                                            <Nav.Link key={"/account"} href={"/account"} style={{"marginLeft": "auto"}}>
-                                                <SmallRedCircle/>
-                                                <strong>Account</strong>
-                                            </Nav.Link>
-                                            <Nav.Link key={"/requestcreation"} href={"/requestcreation"}
-                                                      style={{"marginLeft": "auto"}}>
-                                                <SmallRedCircle/>
-                                                <strong>Create Request</strong>
-                                            </Nav.Link>
-                                            <Nav.Link key={"signout"} href={"#"} onClick={signOut}
-                                                      style={{"marginLeft": "auto"}}>
-                                                <LogoutIcon/>
-                                            </Nav.Link>
-                                        </> :
-                                        <>
-                                            <Nav.Link key={'/signin'} href={'/signin'} style={{"marginLeft": "auto"}}>
-                                                <strong>Sign In</strong>
-                                            </Nav.Link>
-                                            <Nav.Link key={'/signup'} href={'/signup'} style={{"marginLeft": "auto"}}>
-                                                <strong>Sign Up</strong>
-                                            </Nav.Link>
-                                        </>
-                                    }
-                                </Nav>
-                            </Navbar.Collapse>
-                        </Container>
-                    </Navbar>
-                    <main style={{height: `${height - 57}px`}}>
-                        <Routes>
-                            {navLinks.map(({path, component}) => (
-                                <Route key={path} path={path}
-                                       element={React.createElement(component, {token, setToken, role, setRole})}/>
-                            ))}
-                            <Route path="/" element={<Navigate to="/map"/>}/>
-                            <Route path="/rolerequest" state={{token, setToken}}
-                                   element={React.createElement(RoleRequest, {token, setToken})}/>
-                            {
-                                token ? <>
-                                        <Route path="/account" state={{token, setToken}}
-                                               element={React.createElement(Account, {token, setToken})}/>
-                                        <Route path="/requestcreation" state={{token, setToken}}
-                                               element={React.createElement(Request, {token, setToken})}/>
+                                        <Nav.Link key={"/requestcreation"} href={"/requestcreation"}
+                                                  style={{"marginLeft": "auto"}}>
+                                            <SmallRedCircle/>
+                                            <strong>Create Request</strong>
+                                        </Nav.Link>
+                                        <Nav.Link key={"signout"} href={"#"} onClick={signOut}
+                                                  style={{"marginLeft": "auto"}}>
+                                            <LogoutIcon/>
+                                        </Nav.Link>
+                                    </> :
+                                    <>
+                                        <Nav.Link key={'/signin'} href={'/signin'} style={{"marginLeft": "auto"}}>
+                                            <strong>Sign In</strong>
+                                        </Nav.Link>
+                                        <Nav.Link key={'/signup'} href={'/signup'} style={{"marginLeft": "auto"}}>
+                                            <strong>Sign Up</strong>
+                                        </Nav.Link>
                                     </>
-                                    : <>
-                                        <Route path="/signin" state={{token, setToken}}
-                                               element={React.createElement(SignIn, {token, setToken})}/>
-                                        <Route path="/signup" state={{token, setToken}}
-                                               element={React.createElement(SignUp, {token, setToken})}/>
-                                    </>
-                            }
-                        </Routes>
-                    </main>
-                </div>
-            </Router>
-        </LocalizationProvider>
+                                }
+                            </Nav>
+                        </Navbar.Collapse>
+                    </Container>
+                </Navbar>
+                <main style={{height: `${height - 57}px`}}>
+                    <Routes>
+                        {navLinks.map(({path, component}) => (
+                            <Route key={path} path={path}
+                                   element={React.createElement(component, {token, setToken, role, setRole})}/>
+                        ))}
+                        <Route path="/" element={<Navigate to="/map"/>}/>
+                        <Route path="/rolerequest" state={{token, setToken}}
+                               element={React.createElement(RoleRequest, {token, setToken})}/>
+                        {
+                            token ? <>
+                                    <Route path="/account" state={{token, setToken}}
+                                           element={React.createElement(Account, {token, setToken})}/>
+                                    <Route path="/requestcreation" state={{token, setToken}}
+                                           element={React.createElement(Request, {token, setToken})}/>
+                                </>
+                                : <>
+                                    <Route path="/signin" state={{token, setToken}}
+                                           element={React.createElement(SignIn, {token, setToken})}/>
+                                    <Route path="/signup" state={{token, setToken}}
+                                           element={React.createElement(SignUp, {token, setToken})}/>
+                                </>
+                        }
+                    </Routes>
+                </main>
+            </div>
+        </Router>
     );
 }
