+++ resq/frontend/src/components/MapIcons.js
@@ -3,9 +3,9 @@ import * as React from "react";
 
 export const AnnotationIcon = ({icon, color}) =>
     ({
-        Fire: <LocalFireDepartment style={{pointerEvents: "auto", height: "38.6px"}}/>,
-        Health: <LocalHospital style={{pointerEvents: "auto", height: "38.6px", color: color || "#e00000"}}/>,
-        "Road Closure": <Cancel style={{pointerEvents: "auto", height: "38.6px"}}/>
+        fire: <LocalFireDepartment style={{pointerEvents: "auto", height: "38.6px"}}/>,
+        health: <LocalHospital style={{pointerEvents: "auto", height: "38.6px", color: color || "#e00000"}}/>,
+        closed: <Cancel style={{pointerEvents: "auto", height: "38.6px"}}/>
     })[icon]
 export const MarkerIcon = ({color}) => (
     <svg
